services/adk-content-creator/pyproject.toml
[project]
name = "adk-content-creator"
version = "0.1.0"
description = "ADK Agent for creating blog posts and video scripts about Datadog products"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # Google Cloud & AI
    "google-genai>=1.0.0",
    "vertexai>=1.70.0",
    "google-cloud-speech>=2.27.0",
    "google-cloud-storage>=2.18.0",
    
    # Media Processing
    "pillow>=10.4.0",
    "opencv-python>=4.10.0",
    "ffmpeg-python>=0.2.0",
    
    # Content Tools
    "markdown>=3.7",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.9",
    
    # HTTP & Utils
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    
    # Datadog Observability
    "ddtrace>=2.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "black>=25.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "isort>=5.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [{include = "app"}]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "C90", "N", "UP", "YTT", "S", "B", "A", "COM", "DTZ", "T10", "EM", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SIM", "TID", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["S101", "S104", "PLR0913", "I001"]  # Allow assert, binding to 0.0.0.0, many args, disable import checker

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "PLR2004"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"

